snapshot: wip

+++ magick-kt-native-q8/src/commonMain/kotlin/imagemagick/magicknative/NativeChannels.kt
@@ -3,13 +3,14 @@
 package imagemagick.magicknative
 
 import imagemagick.core.enums.Channels
+import platform.posix.size_t
 import kotlin.experimental.ExperimentalNativeApi
 
 @ExperimentalNativeApi
 public object NativeChannels {
-    public inline fun Channels.toNative(): ULong = convertValue(this)
+    public inline fun Channels.toNative(): size_t = convertValue(this)
 
-    public inline fun convertValue(channels: Channels): ULong {
+    public inline fun convertValue(channels: Channels): size_t {
         if (Platform.cpuArchitecture.bitness == 64) {
             return channels.bit
         }
